Add detail page, refactor query page to use detail component, and add uuid link to detail page
